extract osh-interfaces and move implementation to impl.osh
