+++ oshdb-tool/etl/src/main/java/org/heigit/bigspatialdata/oshdb/tool/importer/transform/TransformReader.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.TreeSet;
+import org.heigit.bigspatialdata.oshdb.impl.osh.OSHNodeImpl;
 import org.heigit.bigspatialdata.oshdb.osh.OSHNode;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
 import org.heigit.bigspatialdata.oshdb.tool.importer.osh.TransformOSHWay;
@@ -92,7 +93,7 @@ public class TransformReader{
         int length = data.getInt();
         byte[] content = new byte[length];
         data.get(content);
-        OSHNode node = OSHNode.instance(content, 0, length);
+        OSHNode node = OSHNodeImpl.instance(content, 0, length);
         System.out.println(node.getId());
         ret.add(node);
       }
