+++ oshdb-util/src/main/java/org/heigit/bigspatialdata/oshdb/util/celliterator/CellIterator.java
@@ -178,7 +178,7 @@ public class CellIterator implements Serializable {
       allFullyInside = false;
     }
 
-    Iterable<OSHEntity<OSMEntity>> cellData = (Iterable<OSHEntity<OSMEntity>>) cell;
+    Iterable<? extends OSHEntity> cellData = cell.getEntities();
     return StreamSupport.stream(cellData.spliterator(), false).flatMap(oshEntity -> {
       if (!oshEntityPreFilter.test(oshEntity) ||
           !allFullyInside && (
@@ -189,7 +189,7 @@ public class CellIterator implements Serializable {
         // area of interest -> skip it
         return Stream.empty();
       }
-      if (Streams.stream(oshEntity).noneMatch(osmEntityFilter)) {
+      if (Streams.stream(oshEntity.getVersions()).noneMatch(osmEntityFilter)) {
         // none of this osh entity's versions matches the filter -> skip it
         return Stream.empty();
       }
@@ -442,7 +442,7 @@ public class CellIterator implements Serializable {
     }
 
     //noinspection unchecked
-    Iterable<OSHEntity<OSMEntity>> cellData = (Iterable<OSHEntity<OSMEntity>>) cell;
+    Iterable<? extends OSHEntity> cellData = cell.getEntities();
 
     return StreamSupport.stream(cellData.spliterator(), false).flatMap(oshEntity -> {
       if (!oshEntityPreFilter.test(oshEntity) ||
@@ -454,7 +454,7 @@ public class CellIterator implements Serializable {
         // area of interest -> skip it
         return Stream.empty();
       }
-      if (Streams.stream(oshEntity).noneMatch(osmEntityFilter)) {
+      if (Streams.stream(oshEntity.getVersions()).noneMatch(osmEntityFilter)) {
         // none of this osh entity's versions matches the filter -> skip it
         return Stream.empty();
       }
@@ -523,7 +523,7 @@ public class CellIterator implements Serializable {
                 new LazyEvaluatedObject<>((Geometry)null), prev.geometry,
                 new LazyEvaluatedObject<>((Geometry)null), prev.unclippedGeometry,
                 new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.DELETION)),
-                osmEntity.getChangeset()
+                osmEntity.getChangesetId()
             );
             // cannot normally happen, because prev is never null while skipOutput is true (since no
             // previous result has yet been generated before the first modification in the query
