+++ oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osm/OSMWay.java
@@ -29,7 +29,7 @@ public class OSMWay extends OSMEntity implements Comparable<OSMWay>, Serializabl
 
   public Stream<OSMNode> getRefEntities(OSHDBTimestamp timestamp) {
     return Arrays.stream(this.getRefs()).map(OSMMember::getEntity).filter(Objects::nonNull)
-        .map(entity -> OSHEntities.getByTimestamp((OSHNode) entity, timestamp));
+        .map(entity -> (OSMNode) OSHEntities.getByTimestamp((OSHNode) entity, timestamp));
   }
 
   @Override
